requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
feedparser==6.0.11
google-api-python-client==2.118.0
aiohttp==3.9.3
google-generativeai==0.3.2
pytz==2024.1